style: 개행 축소
