+++ src/test/java/lotto/utils/LottoGeneratorTest.java
@@ -1,7 +1,6 @@
 package lotto.utils;
 
 import lotto.domain.LottoNumber;
-import lotto.domain.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
